+++ R/gfcmSTAR-package.R
@@ -40,6 +40,7 @@
 #' }
 #' \emph{Quality control:}
 #' \tabular{ll}{
+#'   \code{\link{diff.stars}}      \tab show differences between STAR objects\cr
 #'   \code{\link{identical.stars}} \tab check if STAR objects are identical\cr
 #'   \code{\link{peek}}            \tab examine a metadata field\cr
 #'   \code{\link{qc}}              \tab run all quality checks\cr